pyproject.toml
[project]
name = "task4-quiz-generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.0.16",
    "pypdf>=6.3.0",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.1.0",
    "streamlit>=1.45.1",
]
